commit 3

+++ src/main/java/web/dao/UserDao.java
@@ -10,7 +10,7 @@ public interface UserDao {
     void add(User user);
     void delete(User user);
     void edit(User user);
-    User getById(int id);
+    User getById(Long id);
     User fingByUserName(String username);
 
 }
