+++ src/main/java/web/service/UserService.java
@@ -9,6 +9,6 @@ public interface UserService {
     void add(User user);
     void delete(User user);
     void edit(User user);
-    User getById(int id);
+    User getById(Long id);
     User fingByUserName(String username);
 }
